// packages/components/react/src/table/type.ts
export interface TdBaseTableProps<T extends TableRowData = TableRowData> {
  /**
   * 超出省略等所有浮层元素统一绑定到 `attach`，可根据实际情况调整挂载元素
   */
  attach?: AttachNode;
  /**
   * 是否显示表格边框
   * @default false
   */
  bordered?: boolean;
  /**
   * 表格底部内容，可以用于自定义列设置等
   */
  bottomContent?: TNode;
  /**
   * 单元格数据为空时呈现的内容
   */
  cellEmptyContent?: string | TNode<BaseTableCellParams<T>>;
  /**
   * 列配置，泛型 T 指表格数据类型
   * @default []
   */
  columns?: Array<BaseTableCol<T>>;
// ...
}
// ...
export interface BaseTableCol<T extends TableRowData = TableRowData> {
  /**
   * 列横向对齐方式
   * @default left
   */
  align?: 'left' | 'right' | 'center';
  /**
   * 透传 HTML 属性到列元素
   */
  attrs?: BaseTableColumnAttributes<T>;
  /**
   * 自定义单元格渲染，优先级高于 render。泛型 T 指表格数据类型
   */
  cell?: string | TNode<BaseTableCellParams<T>>;
  /**
   * 用于多级表头，泛型 T 指表格数据类型
   */
  children?: Array<BaseTableCol<T>>;
  /**
   * 列类名，值类型是 Function 使用返回值作为列类名；值类型不为 Function 时，值用于整列类名（含表头）。泛型 T 指表格数据类型
   */
  className?: TableColumnClassName<T> | TableColumnClassName<T>[];
  /**
   * 渲染列所需字段，值为 `serial-number` 表示当前列为「序号」列
   * @default ''
   */
  colKey?: string;
  /**
   * 单行表头合并列。多行表头请参考「多级表头」文档示例
   */
  colspan?: number;
  /**
   * 单元格和表头内容超出时，是否显示省略号。如果仅希望单元格超出省略，可设置 `ellipsisTitle = false`。<br/> 值为 `true`，则超出省略浮层默认显示单元格内容；<br/>值类型为 `Function` 则自定义超出省略浮中层显示的内容；<br/>值类型为 `Object`，则自动透传属性到 Tooltip 组件，可用于调整浮层背景色和方向等特性。<br/> 同时透传 Tooltip 属性和自定义浮层内容，请使用 `{ props: { theme: 'light' }, content: () => 'something' }`。<br /> 请注意单元格超出省略的两个基本点：1. 内容元素是内联元素或样式（自定义单元格内容时需特别注意）；2. 内容超出父元素
   * @default false
   */
  ellipsis?:
    | boolean
    | TNode<BaseTableCellParams<T>>
    | TooltipProps
    | { props: TooltipProps; content: TNode<BaseTableCellParams<T>> };
  /**
   * 表头内容超出时，是否显示省略号。优先级高于 `ellipsis`。<br/>值为 `true`，则超出省略的浮层默认显示表头全部内容；<br/>值类型为 `Function` 用于自定义超出省略浮层显示的表头内容；<br/>值类型为 `Object`，则自动透传属性到 Tooltip 组件，则自动透传属性到 Tooltip 组件，可用于调整浮层背景色和方向等特性。<br/> 同时透传 Tooltip 属性和自定义浮层内容，请使用 `{ props: { theme: 'light' }, content: () => 'something' }`
   */
  ellipsisTitle?:
    | boolean
    | TNode<BaseTableColParams<T>>
    | TooltipProps
    | { props: TooltipProps; content: TNode<BaseTableColParams<T>> };
  /**
   * 固定列显示位置
   * @default left
   */
  fixed?: 'left' | 'right';
  /**
   * 自定义表尾内容
   */
  foot?: string | TNode | TNode<{ col: BaseTableCol; colIndex: number }>;
  /**
   * 透传 CSS 属性 `min-width` 到 `<col>` 元素。⚠️ 仅少部分浏览器支持，如：使用 [TablesNG](https://docs.google.com/document/d/16PFD1GtMI9Zgwu0jtPaKZJ75Q2wyZ9EZnVbBacOfiNA/preview) 渲染的 Chrome 浏览器支持 `minWidth`
   */
  minWidth?: string | number;
  /**
   * 自定义表头或单元格，泛型 T 指表格数据类型
   */
  render?: TNode<BaseTableRenderParams<T>>;
  /**
   * 是否允许调整当前列列宽
   * @default true
   */
  resizable?: boolean;
  /**
   * 限制拖拽调整的最小宽度和最大宽度。`resize.minWidth` 默认为 `80`，`resize.maxWidth` 默认为 `600`
   */
  resize?: TableColumnResizeConfig;
  /**
   * 是否阻止当列单元格点击事件冒泡
   */
  stopPropagation?: boolean;
  /**
   * 列表头类名，值类型是函数时使用返回值作为列类名。泛型 T 指表格数据类型
   */
  thClassName?: TableColumnClassName<T> | TableColumnClassName<T>[];
  /**
   * 自定义表头渲染，优先级高于 render
   */
  title?: string | TNode | TNode<{ col: BaseTableCol; colIndex: number }>;
  /**
   * 列宽，可以作为最小宽度使用。当列宽总和小于 `table` 元素时，浏览器根据宽度设置情况自动分配宽度；当列宽总和大于 `table` 元素，表现为定宽。可以同时调整 `table` 元素的宽度来达到自己想要的效果
   */
  width?: string | number;
}
// ...
export interface PrimaryTableCol<T extends TableRowData = TableRowData>
  extends Omit<BaseTableCol, 'cell' | 'title' | 'render' | 'children'> {
  /**
   * 自定义单元格渲染，优先级高于 render。泛型 T 指表格数据类型
   */
  cell?: string | TNode<PrimaryTableCellParams<T>>;
  /**
   * 透传参数，`colKey` 值为 `row-select` 时，配置有效。具体定义参考 Checkbox 组件 和 Radio 组件。泛型 T 指表格数据类型
   */
  checkProps?: CheckProps<T>;
  /**
   * 用于多级表头，泛型 T 指表格数据类型
   */
  children?: Array<PrimaryTableCol<T>>;
  /**
   * 渲染列所需字段，必须唯一。值为 `row-select` 表示当前列为行选中操作列。值为 `drag` 表示当前列为拖拽排序操作列。值为 `serial-number` 表示当前列为「序号」列
   * @default ''
   */
  colKey?: string;
  /**
   * 是否禁用行选中，`colKey` 值为 `row-select` 时，配置有效
   */
  disabled?: (options: { row: T; rowIndex: number }) => boolean;
  /**
   * 可编辑单元格配置项，具体属性参考文档 `TableEditableCellConfig` 描述
   */
  edit?: TableEditableCellConfig<T>;
  /**
   * 过滤规则，支持多选(multiple)、单选(single)、输入框(input) 等三种形式。想要自定义过滤组件，可通过 `filter.component` 实现，自定义过滤组件需要包含参数 value 和事件 change。更多信息请查看当前页面中 `TableColumnFilter` 的详细文档
   */
  filter?: TableColumnFilter;
  /**
   * 自定义表头或单元格，泛型 T 指表格数据类型
   */
  render?: TNode<PrimaryTableRenderParams<T>>;
  /**
   * 当前列支持排序的方式，desc 表示当前列只能进行降序排列；asc 表示当前列只能进行升序排列；all 表示当前列既可升序排列，又可以降序排列
   * @default all
   */
  sortType?: SortType;
  /**
   * 该列是否支持排序。值为 true 表示该列支持排序；值类型为函数，表示对本地数据 `data` 进行排序，返回值参考 [MDN Array.sort](https://developer.mozilla.org/zh-CN/docs/Web/JavaScript/Reference/Global_Objects/Array/sort)。泛型 T 指表格数据类型
   * @default false
   */
  sorter?: boolean | SorterFun<T>;
  /**
   * 自定义表头渲染。值类型为 Function 表示以函数形式渲染表头。值类型为 string 表示使用插槽渲染，插槽名称为 title 的值。优先级高于 render
   */
  title?: string | TNode<{ col: PrimaryTableCol; colIndex: number }>;
  /**
   * `colKey` 值为 `row-select` 时表示行选中列，有两种模式：单选和多选。 `type=single` 表示单选，`type=multiple` 表示多选
   * @default single
   */
  type?: 'single' | 'multiple';
}
// ...
export interface TableColumnFilter {
  /**
   * 用于透传筛选器属性到自定义组件 `component`，HTML 原生属性
   */
  attrs?: HTMLElementAttributes;
  /**
   * 透传类名到自定义组件 `component`
   * @default ''
   */
  classNames?: ClassName;
  /**
   * 用于自定义筛选器，只要保证自定义筛选器包含 value 属性 和 change 事件，即可像内置筛选器一样正常使用。示例：`component: DatePicker`
   */
  component?: ComponentType;
  /**
   * 哪些事件触发后会进行过滤搜索（确认按钮无需配置，会默认触发搜索）。输入框组件示例：`confirmEvents: ['onEnter']`
   */
  confirmEvents?: string[];
  /**
   * 用于配置当前筛选器可选值有哪些，仅当 `filter.type` 等于 `single` 或 `multiple` 时有效
   */
  list?: Array<OptionData>;
  /**
   * 透传 Popup 组件全部属性到筛选器浮层
   */
  popupProps?: PopupProps;
  /**
   * 用于透传筛选器属性到自定义组件 `component`，可以对筛选器进行任何原组件支持的属性配置
   */
  props?: FilterProps;
  /**
   * 重置时设置的值，示例：'' 或 []
   */
  resetValue?: any;
  /**
   * 是否显示重置和确认。值为真，过滤事件（filter-change）会在确定时触发；值为假，则数据变化时会立即触发过滤事件
   * @default false
   */
  showConfirmAndReset?: boolean;
  /**
   * 透传内联样式到自定义组件 `component`
   */
  style?: Styles;
  /**
   * 用于设置筛选器类型：单选按钮筛选器、复选框筛选器、输入框筛选器。更多复杂组件，请更为使用 `component` 自定义任意组件
   * @default ''
   */
  type?: FilterType;
}
// ...
export interface TableEditableCellConfig<T extends TableRowData = TableRowData> {
  /**
   * 除了点击非自身元素退出编辑态之外，还有哪些事件退出编辑态。示例：`abortEditOnEvent: ['onChange']`
   */
  abortEditOnEvent?: string[];
  /**
   * 组件定义，如：`Input` `Select`。对于完全自定义的组件（非组件库内的组件），组件需要支持 `value` 和 `onChange` ；如果还需要支持校验规则，则组件还需实现 `tips` 和 `status` 两个 API，实现规则可参考 `Input` 组件
   */
  component?: ComponentType;
  /**
   * 单元格默认状态是否为编辑态
   * @default false
   */
  defaultEditable?: boolean;
  /**
   * 设置当前列的单元格始终保持为编辑态
   * @default false
   */
  keepEditMode?: boolean;
  /**
   * 透传给编辑组件的事件
   */
  on?: (context: TableEditableCellPropsParams<T>) => { [eventName: string]: Function };
  /**
   * 编辑完成后，退出编辑模式时触发
   */
  onEdited?: (context: PrimaryTableOnEditedContext<T>) => void;
  /**
   * 透传给组件 `edit.component` 的属性
   */
  props?: TableEditableCellProps<T>;
  /**
   * 校验规则
   */
  rules?: TableEditableCellRules<T>;
  /**
   * 是否显示编辑图标
   * @default true
   */
  showEditIcon?: boolean;
  /**
   * 触发校验的时机，有 2 种：退出编辑时和数据变化时
   * @default 'exit'
   */
  validateTrigger?: 'exit' | 'change';
}
// ...
export interface TableRowData {
  [key: string]: any;
  children?: TableRowData[];
}

export type BaseTableColumnAttributes<T> = { [key: string]: any } | ((context: CellData<T>) => { [key: string]: any });

export interface BaseTableCellParams<T> {
  row: T;
  rowIndex: number;
  col: BaseTableCol<T>;
  colIndex: number;
}

export type TableColumnClassName<T> = ClassName | ((context: CellData<T>) => ClassName);

export interface CellData<T> extends BaseTableCellParams<T> {
  type: 'th' | 'td';
}

export interface BaseTableColParams<T> {
  col: BaseTableCol<T>;
  colIndex: number;
}

export interface BaseTableRenderParams<T> extends BaseTableCellParams<T> {
  type: RenderType;
}

export type RenderType = 'cell' | 'title';

export interface TableColumnResizeConfig {
  minWidth: number;
  maxWidth: number;
}
// ...
export interface PrimaryTableCellParams<T> {
  row: T;
  rowIndex: number;
  col: PrimaryTableCol<T>;
  colIndex: number;
}

export type CheckProps<T> =
  | CheckboxProps
  | RadioProps
  | ((options: { row: T; rowIndex: number }) => CheckboxProps | RadioProps);

export interface PrimaryTableRenderParams<T> extends PrimaryTableCellParams<T> {
  type: RenderType;
}

export type SortType = 'desc' | 'asc' | 'all';

export type SorterFun<T> = (a: T, b: T) => number;
// ...
export type FilterProps = RadioProps | CheckboxProps | InputProps | { [key: string]: any };

export type FilterType = 'input' | 'single' | 'multiple';

export type PrimaryTableOnEditedContext<T> = PrimaryTableCellParams<T> & { trigger: string; newRowData: T };

export type TableEditableCellProps<T> =
  | TablePlainObject
  | ((params: TableEditableCellPropsParams<T>) => TablePlainObject);

export interface TableEditableCellPropsParams<T> extends PrimaryTableCellParams<T> {
  editedRow: T;
  updateEditedCellValue: (val: any) => void;
}

export interface TablePlainObject {
  [key: string]: any;
}

export type TableEditableCellRules<T> = FormRule[] | ((params: PrimaryTableCellParams<T>) => FormRule[]);